DataDictionary Changes Data Type for demographic languages

+++ EDI/ApplicationCore/Entities/QuestionnairesDataDemographic.cs
@@ -44,9 +44,9 @@ namespace EDI.ApplicationCore.Entities
         public byte? OtherLanguage { get; set; }
         public byte? IndigenousAncestry { get; set; }
         public byte? AfricanDescent { get; set; }
-        public byte? FirstLanguage1 { get; set; }
-        public byte? FirstLanguage2 { get; set; }
-        public byte? FirstLanguage3 { get; set; }
+        public int? FirstLanguage1 { get; set; }
+        public int? FirstLanguage2 { get; set; }
+        public int? FirstLanguage3 { get; set; }
         public byte? LandAgreement { get; set; }
         [JsonIgnore]
         public virtual Child Child { get; set; }
